requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
PyMySQL
PyJWT>=2.9.0